agents/chatbot/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
streamlit>=1.30.0
python-dotenv>=1.0.0
